# docs/api/api_definition.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ContentFlow Enterprise API",
    "description": "API completa para la integración programática con ContentFlow. Disponible exclusivamente para clientes del plan Enterprise.",
    "version": "1.0.0",
    "contact": {
      "email": "enterprise-support@contentflow.app"
    }
  },
  "servers": [
    {
      "url": "https://tu-dominio.com/api/v1",
      "description": "Servidor de Producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/publications": {
      "get": {
        "tags": ["Publicaciones"],
        "summary": "Listar publicaciones del workspace activo",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" }, "description": "Filtrar por estado(s) separados por coma. Ej: draft,approved" },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "platform", "in": "query", "schema": { "type": "string" } },
          { "name": "date_start", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "date_end", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["newest","oldest","title_asc","title_desc"] } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "simplified", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "Lista paginada de publicaciones" }
        }
      },
      "post": {
        "tags": ["Publicaciones"],
        "summary": "Crear una nueva publicación",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "body"],
                "properties": {
                  "title": { "type": "string" },
                  "body": { "type": "string" },
                  "description": { "type": "string" },
                  "hashtags": { "type": "string" },
                  "url": { "type": "string", "format": "uri" },
                  "goal": { "type": "string", "enum": ["brand_awareness","lead_generation","engagement","sales"] },
                  "status": { "type": "string", "enum": ["draft","scheduled"], "default": "draft" },
                  "scheduled_at": { "type": "string", "format": "date-time" },
                  "platform_settings": { "type": "object" },
                  "media": { "type": "array", "items": { "type": "object" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Publicación creada" },
          "402": { "description": "Límite del plan alcanzado" },
          "403": { "description": "Sin permisos" }
        }
      }
    },
...
  }
}